+++ src/RcppExports.cpp
@@ -36,6 +36,17 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// bvartvpalg_kalman_dk
+Rcpp::List bvartvpalg_kalman_dk(Rcpp::List object);
+RcppExport SEXP _bvartools_bvartvpalg_kalman_dk(SEXP objectSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::List >::type object(objectSEXP);
+    rcpp_result_gen = Rcpp::wrap(bvartvpalg_kalman_dk(object));
+    return rcpp_result_gen;
+END_RCPP
+}
 // bvecalg
 Rcpp::List bvecalg(Rcpp::List object);
 RcppExport SEXP _bvartools_bvecalg(SEXP objectSEXP) {
@@ -487,6 +498,7 @@ RcppExport SEXP _bvartools_RcppExport_registerCCallable() {
 static const R_CallMethodDef CallEntries[] = {
     {"_bvartools_bvaralg", (DL_FUNC) &_bvartools_bvaralg, 1},
     {"_bvartools_bvartvpalg", (DL_FUNC) &_bvartools_bvartvpalg, 1},
+    {"_bvartools_bvartvpalg_kalman_dk", (DL_FUNC) &_bvartools_bvartvpalg_kalman_dk, 1},
     {"_bvartools_bvecalg", (DL_FUNC) &_bvartools_bvecalg, 1},
     {"_bvartools_bvectvpalg", (DL_FUNC) &_bvartools_bvectvpalg, 1},
     {"_bvartools_bvs", (DL_FUNC) &_bvartools_bvs, 7},